+++ src/Mappers/Loyalty/Rewards/DigitalRewardMapper.php
@@ -26,6 +26,8 @@ class DigitalRewardMapper
 
         $active = property_exists($data, 'active') ? $data->active : true;
 
+        $attributesNamesToDelete = ['uuid', 'title', 'description', 'required_credits', 'reward_type', 'media', 'active', 'is_active', 'id', 'stock', 'cost_price'];
+        $attributes = array_diff_key(get_object_vars($data), array_flip($attributesNamesToDelete));
 
         return new DigitalReward(
             $data->uuid,
@@ -34,8 +36,8 @@ class DigitalRewardMapper
             $media ?? null,
             $data->description ?? "",
             $active,
-            RewardType::byName($data->reward_type)->getValue() ?? null
-
+            RewardType::byName($data->reward_type)->getValue() ?? null,
+            $attributes
         );
     }
 }