+++ src/Mappers/Loyalty/Rewards/PhysicalRewardMapper.php
@@ -26,6 +26,9 @@ class PhysicalRewardMapper
 
         $active = property_exists($data, 'active') ? $data->active : true;
 
+        $attributesNamesToDelete = ['uuid', 'title', 'description', 'required_credits', 'reward_type', 'media', 'active', 'is_active', 'id', 'stock', 'cost_price'];
+        $attributes = array_diff_key(get_object_vars($data), array_flip($attributesNamesToDelete));
+
         return new PhysicalReward(
             $data->uuid,
             $data->title ?? '',
@@ -33,7 +36,8 @@ class PhysicalRewardMapper
             $media ?? null,
             $data->description ?? "",
             $active,
-            RewardType::byName($data->reward_type)->getValue() ?? null
+            RewardType::byName($data->reward_type)->getValue() ?? null,
+            $attributes
         );
     }
 }