+++ src/Models/Loyalty/Rewards/PhysicalReward.php
@@ -11,9 +11,9 @@ use Piggy\Api\Models\Loyalty\Media;
 class PhysicalReward extends Reward
 {
 
-    public function __construct(string $uuid, ?string $title = "", ?int $requiredCredits = null, ?Media $media = null, ?string $description = "", ?bool $active = true, ?string $rewardType = null)
+    public function __construct(string $uuid, ?string $title = "", ?int $requiredCredits = null, ?Media $media = null, ?string $description = "", ?bool $active = true, ?string $rewardType = null, array $attributes = [])
     {
-        parent::__construct($uuid, $title, $requiredCredits, $media, $description, $active, $rewardType);
+        parent::__construct($uuid, $title, $requiredCredits, $media, $description, $active, $rewardType, $attributes);
     }
 
 }