+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 3.0.4
 - feat: option to get Loyalty Program (and see settings)
+- feat: update Reward calls for Register and OAuth Client added. Can update custom Reward Attributes too.
 
 ## 3.0.3
 - feat: option to send loyalty transaction custom attributes with the create credit reception requests for both oauth and register client